add sort and title fucntion and moffy api link to prevent the filter
